most of rays comments

+++ DbTables/inc/CalEnergyCalib.hh
@@ -56,7 +56,7 @@ namespace mu2e {
     if (index!=int(_rows.size())) {
         throw cet::exception("CALOENERGYCALIB_BAD_INDEX")<<"CalEnergyCalib::addRow found index out of order:"<<index << " != " << int(_rows.size()) <<"\n";
       }
-       _rows.emplace_back(CaloSiPMId(index),std::stof(columns[1]),std::stof(columns[2]),std::stof(columns[3]));
+       _rows.emplace_back(CaloSiPMId(index),std::stof(columns[1]),std::stof(columns[2]),std::stoi(columns[3]));
 
     }
